Fixing hyprctl notify color issue
hyprctl`s notify service has a different color expression logic. Colors can be Rgb or Rgba and the value of them has to be written in hex format. former system wrote colors as like as this: rgb(25, 255, 100). however, the correct form has to be like this: rgb(ff,e5,aa).